+++ variables.tf
@@ -63,6 +63,12 @@ variable "vm_size" {
   type        = string
 }
 
+variable "vm_zone" {
+  description = "Bastion virtual machine zone"
+  type        = number
+  default     = 1
+}
+
 variable "admin_username" {
   description = "Name of the admin user"
   type        = string
@@ -117,5 +123,24 @@ variable "storage_os_disk_caching" {
 
 variable "storage_os_disk_size_gb" {
   description = "Specifies the size of the OS Disk in gigabytes"
+  type        = string
+}
+
+## Identity variables
+variable "identity" {
+  description = "Map with identity block informations as described here https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/linux_virtual_machine#identity"
+  type = object({
+    type         = string
+    identity_ids = list(string)
+  })
+  default = {
+    type         = "SystemAssigned"
+    identity_ids = []
+  }
+}
+
+## Backup variable
+variable "backup_policy_id" {
+  description = "Backup policy ID from the Recovery Vault to attach the Virtual Machine to (value to `null` to disable backup)"
   type        = string
 }